+++ rgen/rgen-placer/src/placer/tree/dead_tree.rs
@@ -15,11 +15,11 @@ impl Placer for DeadTree {
   fn place(&self, world: &mut PartialWorld, rng: &mut Rng, pos: Pos) {
     let height = rng.rand_inclusive(4, 7);
 
-    if pos.y as i32 + height as i32 + 2 >= 255 || pos.y <= 1 {
+    if pos.y + height + 2 >= 255 || pos.y <= 1 {
       return;
     }
 
-    for y in 0..height as u8 {
+    for y in 0..height {
       world.set(pos + Pos::new(0, y, 0), self.trunk);
     }
   }
